+++ src/InputWrapper.js
@@ -11,26 +11,49 @@ class InputWrapper extends Component {
      * https://github.com/facebook/react/issues/4999#issuecomment-144557619
      */
     componentDidMount() {
-        const { onEarlyInput, value, checked } = this.props;
+        const { onChange, onEarlyInput, value, checked } = this.props;
 
         if (
             this.inputNode !== undefined &&
-            typeof onEarlyInput === 'function'
+            (typeof onEarlyInput === 'function' ||
+                typeof onChange === 'function')
         ) {
             const type = this.inputNode.type;
 
-            if (
-                (type === 'checkbox' || type === 'radio') &&
-                this.inputNode.checked !== checked
-            ) {
+            /**
+             * Checks if values are out of sync
+             */
+            if (type === 'checkbox' || type === 'radio') {
+                if (this.inputNode.checked === checked) {
+                    return;
+                }
+            } else if (this.inputNode.value === value) {
+                return;
+            }
+
+            /**
+             * Only call onEarlyInput if specified
+             *
+             * Used to handle custom events not wanted with typical onChange, such as 'touched'.
+             */
+            if (onEarlyInput) {
                 onEarlyInput(this.inputNode);
 
                 return;
             }
 
-            if (this.inputNode.value !== value) {
-                onEarlyInput(this.inputNode);
-            }
+            /**
+             * Simulate onChange event. Send PR for missing needed props
+             */
+            const event = {
+                preventDefault() {
+                    return null;
+                },
+
+                target: this.inputNode,
+            };
+
+            onChange(event);
         }
     }
 
@@ -44,13 +67,15 @@ class InputWrapper extends Component {
 
         const value = this.props.value;
         const checked = this.props.checked;
+        const onChange = this.props.onChange;
 
         return (
             <HTMLInputTag
                 ref={input => {
                     if (
                         input &&
-                        typeof onEarlyInput === 'function' &&
+                        (typeof onEarlyInput === 'function' ||
+                            typeof onChange === 'function') &&
                         (value !== undefined || checked !== undefined) &&
                         this.inputNode === undefined
                     ) {